+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.1] - 2025-03-10
+
+- Update deps.
+- Clarify docs on supported Zig versions.
+
 ## [0.6.0] - 2025-03-10
 
 - Bump default Zig version to `0.14.0`.